+++ .idea/misc.xml
@@ -4,7 +4,6 @@
     <option name="originalFiles">
       <list>
         <option value="$PROJECT_DIR$/pom.xml" />
-        <option value="C:/Try/SimpleRegistrationExample/pom.xml" />
       </list>
     </option>
   </component>
