+++ src/main/java/com/example/delivery_of_services_application/controllers/LoginController.java
@@ -36,12 +36,7 @@ public class LoginController {
     private AnchorPane openClientInterface;
     @FXML
     private AnchorPane openInstructorInterface;
-    @FXML
-    private ChoiceBox nextTo;
 
-    public void initialize(){
-        nextTo.getItems().addAll("Customer","Provider");
-    }
     @FXML
     public void loginButtonOnAction(ActionEvent event) throws IOException{
 
@@ -58,7 +53,7 @@ public class LoginController {
                         Parent root;
                         stage = (Stage) loginButton.getScene().getWindow();
                         root=FXMLLoader.load(getClass().getClassLoader().getResource("ProviderHomePage.fxml"));
-                        stage.setScene(new Scene(root, 1127, 680));
+                        stage.setScene(new Scene(root, 800, 480));
                         stage.show();
                     }
                     if(k==2){
@@ -66,8 +61,8 @@ public class LoginController {
                         Stage stage;
                         Parent root;
                         stage = (Stage) loginButton.getScene().getWindow();
-                        root=FXMLLoader.load(getClass().getClassLoader().getResource(" CustomerHomePage.fxml"));
-                        stage.setScene(new Scene(root, 1127, 680));
+                        root=FXMLLoader.load(getClass().getClassLoader().getResource("CustomerHomePage.fxml"));
+                        stage.setScene(new Scene(root, 800, 480));
                         stage.show();
                     }
