+++ src/main/java/com/example/delivery_of_services_application/controllers/ProviderHomePageController.java
@@ -1,0 +1,4 @@
+package com.example.delivery_of_services_application.controllers;
+
+public class ProviderHomePageController {
+}
