--- src/main/java/com/example/delivery_of_services_application/controllers/provider/ProviderHomePageController.java
@@ -1,4 +1,0 @@
-package com.example.delivery_of_services_application.controllers.provider;
-
-public class ProviderHomePageController {
-}
