+++ src/main/resources/CustomerHomePage.fxml
@@ -1,54 +1,78 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.scene.control.*?>
+<?import javafx.scene.effect.*?>
 <?import javafx.scene.image.*?>
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 
-<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="490.0" prefWidth="660.0" xmlns="http://javafx.com/javafx/11.0.14-internal" xmlns:fx="http://javafx.com/fxml/1"
-fx:controller="com.example.delivery_of_services_application.controllers.customer.CustomerHomePageController">
-   <left>
-      <AnchorPane prefHeight="490.0" prefWidth="240.0" style="-fx-background-color: #0B4C6C;" BorderPane.alignment="CENTER">
+<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="490.0" prefWidth="660.0" xmlns="http://javafx.com/javafx/11.0.14-internal" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.delivery_of_services_application.controllers.CustomerHomePageController">
+    <left>
+        <AnchorPane prefHeight="490.0" prefWidth="156.0" style="-fx-background-color: #0B4C6C;" BorderPane.alignment="CENTER">
+            <children>
+                <ImageView fitHeight="131.0" fitWidth="125.0" layoutX="14.0" layoutY="29.0" pickOnBounds="true" preserveRatio="true">
+                    <image>
+                        <Image url="@images/homeC.png" />
+                    </image>
+                </ImageView>
+                <Button layoutX="17.0" layoutY="210.0" mnemonicParsing="false" prefHeight="25.0" prefWidth="119.0" text="Profile">
+               <effect>
+                  <InnerShadow />
+               </effect></Button>
+                <Button layoutX="17.0" layoutY="274.0" mnemonicParsing="false" prefHeight="25.0" prefWidth="120.0" text="History">
+               <effect>
+                  <InnerShadow />
+               </effect></Button>
+                <Button layoutX="16.0" layoutY="332.0" mnemonicParsing="false" onAction="#backButtonOnAction" prefHeight="25.0" prefWidth="121.0" text="Log Out">
+               <effect>
+                  <InnerShadow />
+               </effect></Button>
+            </children>
+        </AnchorPane>
+    </left>
+    <right>
+        <AnchorPane prefHeight="490.0" prefWidth="144.0" style="-fx-background-color: #A1BEBF;" BorderPane.alignment="CENTER">
          <children>
-            <ImageView fitHeight="150.0" fitWidth="200.0" layoutX="45.0" layoutY="20.0" pickOnBounds="true" preserveRatio="true">
-               <image>
-                  <Image url="@images/homeC.png" />
-               </image>
-            </ImageView>
-            <Button layoutX="25.0" layoutY="210.0" mnemonicParsing="false" prefHeight="25.0" prefWidth="197.0" text="Profile" />
-            <Button layoutX="25.0" layoutY="266.0" mnemonicParsing="false" prefHeight="25.0" prefWidth="197.0" text="History" />
-            <Button fx:id="logOutButton" layoutX="25.0" layoutY="321.0" mnemonicParsing="false" onAction="#backToLogin" prefHeight="25.0" prefWidth="197.0" text="Log Out" />
-         </children>
-      </AnchorPane>
-   </left>
-   <right>
-      <AnchorPane prefHeight="490.0" prefWidth="426.0" style="-fx-background-color: #A1BEBF;" BorderPane.alignment="CENTER">
-         <children>
-            <Button layoutX="11.0" layoutY="132.0" mnemonicParsing="false" text="Agricultura" />
-            <Button layoutX="15.0" layoutY="172.0" mnemonicParsing="false" text="Curatenie" />
-            <Button layoutX="15.0" layoutY="273.0" mnemonicParsing="false" text="Hardware" />
-            <Button layoutX="121.0" layoutY="273.0" mnemonicParsing="false" text="Instalatii" />
-            <Button layoutX="112.0" layoutY="356.0" mnemonicParsing="false" text="Software" />
-            <Button layoutX="24.0" layoutY="318.0" mnemonicParsing="false" text="Juridic" />
-            <Button layoutX="112.0" layoutY="318.0" mnemonicParsing="false" text="Mancare" />
-            <Button layoutX="112.0" layoutY="132.0" mnemonicParsing="false" text="Automobile" />
-            <Button layoutX="116.0" layoutY="172.0" mnemonicParsing="false" text="Constructii" />
-            <Button layoutX="14.0" layoutY="220.0" mnemonicParsing="false" text="Electronice" />
-            <Button layoutX="16.0" layoutY="398.0" mnemonicParsing="false" text="Traduceri" />
-            <Button layoutX="121.0" layoutY="220.0" mnemonicParsing="false" text="Educatie" />
-            <Label layoutX="16.0" layoutY="65.0" prefHeight="17.0" prefWidth="227.0" text="Domains">
+            <Button layoutX="13.0" layoutY="24.0" mnemonicParsing="false" onAction="#cleaningButtonOnAction" prefHeight="51.0" prefWidth="120.0" text="Cleaning">
                <font>
-                  <Font size="27.0" />
+                  <Font name="System Italic" size="12.0" />
                </font>
-            </Label>
-            <Button layoutX="202.0" layoutY="220.0" mnemonicParsing="false" text="Evenimente" />
-            <Button layoutX="204.0" layoutY="273.0" mnemonicParsing="false" text="Transpoort" />
-            <Button layoutX="15.0" layoutY="356.0" mnemonicParsing="false" text="Publicitate" />
-            <Button layoutX="208.0" layoutY="166.0" mnemonicParsing="false" text="Consiliere" />
-            <Button layoutX="219.0" layoutY="356.0" mnemonicParsing="false" text="Sport" />
-            <Button layoutX="213.0" layoutY="318.0" mnemonicParsing="false" text="Nutritie" />
-            <Button layoutX="116.0" layoutY="398.0" mnemonicParsing="false" text="Vanzari" />
+               <effect>
+                  <SepiaTone />
+               </effect>
+            </Button>
+            <Button layoutX="14.0" layoutY="108.0" mnemonicParsing="false" onAction="#renovationWorkButtonOnAction" prefHeight="62.0" prefWidth="118.0" text="Renovation Work">
+               <font>
+                  <Font name="System Italic" size="12.0" />
+               </font>
+               <effect>
+                  <SepiaTone />
+               </effect></Button>
+            <Button layoutX="14.0" layoutY="203.0" mnemonicParsing="false" onAction="#medicalServicesButtonOnAction" prefHeight="59.0" prefWidth="118.0" text="Medical Services">
+               <font>
+                  <Font name="System Italic" size="12.0" />
+               </font>
+               <effect>
+                  <SepiaTone />
+               </effect></Button>
+            <Button layoutX="17.0" layoutY="294.0" mnemonicParsing="false" onAction="#itSupportButtonOnAction" prefHeight="54.0" prefWidth="114.0" text="IT Support">
+               <font>
+                  <Font name="System Italic" size="12.0" />
+               </font>
+               <effect>
+                  <SepiaTone />
+               </effect></Button>
+            <Button layoutX="17.0" layoutY="381.0" mnemonicParsing="false" onAction="#othersButtonOnAction" prefHeight="58.0" prefWidth="112.0" text="Others...">
+               <effect>
+                  <SepiaTone />
+               </effect>
+               <font>
+                  <Font name="System Italic" size="12.0" />
+               </font></Button>
          </children>
-      </AnchorPane>
-   </right>
+        </AnchorPane>
+    </right>
+   <center>
+      <AnchorPane prefHeight="200.0" prefWidth="200.0" BorderPane.alignment="CENTER" />
+   </center>
 </BorderPane>
